Split out XdgPortalRequest
This is so the portals can be split out without exposing too much internal API.
